+++ docs/changelog.md
@@ -10,12 +10,19 @@ See the [GitHub "Features" project](https://github.com/users/peter88213/projects
 
 ---
 
+### Version 5.46.7
+
+- Fixed tag selection list display.
+- Refactored CollectionBox, using a `ttk.Treeview` list. 
+  This allows to enable the buttons on element selection.
+- Refactored `PickList`, simplifying the code.
+
+
 ### Version 5.46.6
 
 - Changed the "Settings" icon.
 - Refactored the code for better maintainability, replacing the classes `DataImportDialog`
   and `StrSelectionDialog` with a generic and improved `PickList` class.
-
 
 
 ### Version 5.46.5